+++ CHANGELOG.md
@@ -32,13 +32,13 @@ For more information, see [UP42 Python package description](https://pypi.org/pro
 
 ## 2.1.0a2
 
-**TBD, 2024**
+**Sep 2, 2024**
 - Fix test coverage for `Order:estimate`, `Order:place`, and `Order:track_status` methods.
 - change `Order.track_status` report_time input type to float.
 
 ## 2.1.0a1
 
-**TBD, 2024**
+**Aug 30, 2024**
 - Improve test coverage `Order` class.
 
 ## 2.0.1