+++ CHANGELOG.md
@@ -34,7 +34,7 @@ For more information, see [UP42 Python package description](https://pypi.org/pro
 
 **Sep 2, 2024**
 - Fix test coverage for `Order:estimate`, `Order:place`, and `Order:track_status` methods.
-- change `Order.track_status` report_time input type to float.
+- Change `Order:track_status` report_time input type to float.
 
 ## 2.1.0a1