+++ CHANGELOG.md
@@ -33,8 +33,8 @@ For more information, see [UP42 Python package description](https://pypi.org/pro
 ## 2.1.0a2
 
 **Sep 2, 2024**
-- Fix test coverage for `Order:estimate`, `Order:place`, and `Order:track_status` methods.
-- Change `Order:track_status` report_time input type to float.
+- Fix test coverage for `Order::estimate`, `Order::place`, and `Order::track_status` methods.
+- Change `Order::track_status` report_time input type to float.
 
 ## 2.1.0a1